// src/interfaces/Posts.tsx
export interface PostCategory {
    id:                     number;
    title:                  string;
    excerpt:                string;
    date:                   string;
    content:                string;
    slug:                   string;
    guid:                   string;
    author:                 string;
    category:               Category[];
    tema:                   any[];
    tema_escaner:           boolean;
    ruta_geografica:        PostTag[];
    featured_image:         FeaturedImage;
    post_tags:              PostTag[];
    autores_escaner:        boolean;
    Youtube_URL:            string;
    mostrar_firma:          boolean;
    colaboradores_escaners: null;
}

export interface Category {
    term_id:              number;
    name:                 string;
    slug:                 string;
    term_group:           number;
    term_taxonomy_id:     number;
    taxonomy:             string;
    description:          string;
    parent:               number;
    count:                number;
    filter:               string;
    cat_ID:               number;
    category_count:       number;
    category_description: string;
    cat_name:             string;
    category_nicename:    string;
    category_parent:      number;
}

export interface FeaturedImage {
    thumbnail: string;
    medium:    string;
    large:     string;
}

export interface PostTag {
    term_id:          number;
    name:             string;
    slug:             string;
    term_group:       number;
    term_taxonomy_id: number;
    taxonomy:         string;
    description:      string;
    parent:           number;
    count:            number;
    filter:           string;
}